Extract Listable concern and include in ShoppingListItem model
